# src/server/routes/v1/openapi.yaml
openapi: 3.0.3
info:
  title: MeshMonitor API
  description: |
    MeshMonitor REST API v1 provides programmatic access to mesh network data.

    ## Multi-source (breaking change — spec 2.0.0)
    MeshMonitor 4.0 runs one installation against many Meshtastic/MQTT/MeshCore
    sources. Per-resource endpoints are now canonical under
    `/api/v1/sources/{sourceId}/...`:

    | Resource | Canonical path |
    |---|---|
    | Nodes | `/api/v1/sources/{sourceId}/nodes` |
    | Messages | `/api/v1/sources/{sourceId}/messages` |
    | Channels | `/api/v1/sources/{sourceId}/channels` |
    | Telemetry | `/api/v1/sources/{sourceId}/telemetry` |
    | Traceroutes | `/api/v1/sources/{sourceId}/traceroutes` |
    | Packets | `/api/v1/sources/{sourceId}/packets` |
    | Network | `/api/v1/sources/{sourceId}/network` |
    | Position history | `/api/v1/sources/{sourceId}/nodes/{nodeId}/position-history` |
    | Status | `/api/v1/sources/{sourceId}/status` |

    The literal string `default` may be used for `{sourceId}` — it resolves to
    the first source the authenticated token has `read` permission on (or the
    first enabled source for admin tokens). 404 if none is accessible.

    `GET /api/v1/sources` lists the sources the token can read.

    `/api/v1/solar` and `/api/v1/channel-database` remain deployment-global
    (no source scope).

    Legacy root paths (`/api/v1/nodes?sourceId=...`, etc.) were removed in
...
    Use the per-source shape instead.

    ## Authentication
...

    ```
...

...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Token
      description: |
        API token authentication. Use format: `Bearer mm_v1_<token>`

        Generate tokens from your User Settings page.
  parameters:
    SourceIdParam:
      name: sourceId
      in: path
      required: true
      description: |
        Source identifier. Either a UUID returned by `GET /api/v1/sources`, or
        the literal string `default` to resolve the first source the
        authenticated token has `read` permission on (admins get the first
        enabled source). Returns 404 if no accessible source exists.
      schema:
        type: string
        example: default
  schemas:
    Error:
      type: object
      required:
        - success
        - error
        - message
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Unauthorized"
        message:
          type: string
          example: "API token required"
...
    Node:
      type: object
      properties:
        nodeId:
          type: string
          description: Unique node identifier (hex format)
          example: "!a1b2c3d4"
        nodeNum:
          type: integer
          description: Node number
          example: 2715451348
        longName:
          type: string
          description: Long name of the node
          example: "My Mesh Node"
        shortName:
          type: string
          description: Short name (4 characters)
          example: "MESH"
        role:
          type: string
          description: Node role in the mesh network
          enum: [CLIENT, ROUTER, ROUTER_CLIENT]
          example: "CLIENT"
        hwModel:
          type: string
          description: Hardware model
          example: "TBEAM"
        latitude:
          type: number
          format: double
          description: Node latitude
          example: 37.7749
        longitude:
          type: number
          format: double
          description: Node longitude
          example: -122.4194
        altitude:
          type: integer
          description: Node altitude in meters
          example: 50
        hopsAway:
          type: integer
          description: Number of hops from primary node
          example: 1
        lastHeard:
          type: integer
          description: Unix timestamp of last contact
          example: 1699564800
        snr:
          type: number
          format: float
          description: Signal-to-noise ratio
          example: 8.5
...
    NodeResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Node'
...
    Packet:
      type: object
      properties:
        id:
          type: integer
          description: Unique packet log ID
          example: 12345
        packet_id:
          type: integer
          description: Packet identifier
          example: 987654
        timestamp:
          type: integer
          description: Unix timestamp
          example: 1699564800
        from_node:
          type: integer
          description: Source node number
          example: 2715451348
        from_node_id:
          type: string
          description: Source node ID (hex format)
          example: "!a1b2c3d4"
        from_node_longName:
          type: string
          description: Source node long name
          example: "My Mesh Node"
        to_node:
          type: integer
          description: Destination node number
          example: 2820214276
        to_node_id:
          type: string
          description: Destination node ID (hex format)
          example: "!e5f6g7h8"
        to_node_longName:
          type: string
          description: Destination node long name
          example: "Remote Node"
        channel:
          type: integer
          description: Channel number
          example: 0
        portnum:
          type: integer
          description: Port number (message type)
          example: 1
        portnum_name:
          type: string
          description: Port name
          example: "TEXT_MESSAGE_APP"
        encrypted:
          type: boolean
          description: Whether packet was encrypted
          example: false
        snr:
          type: number
          format: float
          description: Signal-to-noise ratio
          example: 8.5
        rssi:
          type: integer
          description: Received signal strength in dBm
          example: -95
        hop_limit:
          type: integer
          description: Hop limit
          example: 3
        hop_start:
          type: integer
          description: Starting hop count
          example: 3
        payload_size:
          type: integer
          description: Size of payload in bytes
          example: 128
        want_ack:
          type: boolean
          description: Whether acknowledgment was requested
          example: false
        priority:
          type: integer
          description: Packet priority
          example: 64
    PacketListResponse:
      type: object
      required:
        - success
        - count
        - total
        - offset
        - limit
        - data
      properties:
        success:
          type: boolean
          example: true
        count:
          type: integer
          description: Number of packets returned in this response
          example: 100
        total:
          type: integer
          description: Total number of packets matching the filters
          example: 5432
        offset:
          type: integer
          description: Offset used for pagination
          example: 0
        limit:
          type: integer
          description: Limit used for pagination
          example: 100
        data:
          type: array
          items:
            $ref: '#/components/schemas/Packet'
    PacketResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Packet'
...
paths:
...
  /sources/{sourceId}/nodes:
    parameters:
      - $ref: '#/components/parameters/SourceIdParam'
    get:
      tags:
        - Nodes
      summary: Get nodes for a source (canonical)
      description: |
        Returns nodes visible for the given source, filtered by the token's
        per-source channel read permissions. This is the canonical replacement
        for the removed root-level `GET /api/v1/nodes?sourceId=...` shape.
      parameters:
        - name: active
          in: query
          description: Filter to only nodes active within the last `sinceDays` days
          required: false
          schema:
            type: boolean
            default: false
        - name: sinceDays
          in: query
          description: Activity window in days
          required: false
          schema:
            type: integer
            default: 7
            minimum: 1
      responses:
        '200':
          description: Nodes retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Node'
        '403':
          description: Insufficient permissions on the source
        '404':
          description: Source not found
  /nodes/{nodeId}:
    get:
      tags:
        - Nodes
      summary: Get a specific node
      description: Retrieve detailed information about a specific node
      parameters:
        - name: nodeId
          in: path
          description: Node ID (hex format, e.g., !a1b2c3d4)
          required: true
          schema:
            type: string
            pattern: '^![0-9a-fA-F]{8}$'
      responses:
        '200':
          description: Node retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeResponse'
        '404':
          description: Node not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /packets:
    get:
      tags:
        - Packets
      summary: Get packet logs
      description: |
        Retrieve raw packet log data from the mesh network with optional filters.
        This endpoint provides access to all packets received by the node, including both encrypted and decrypted packets.
      parameters:
        - name: offset
          in: query
          description: Number of packets to skip for pagination
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          description: Maximum number of packets to return
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
        - name: portnum
          in: query
          description: Filter by port number (message type)
          required: false
          schema:
            type: integer
        - name: from_node
          in: query
          description: Filter by source node number
          required: false
          schema:
            type: integer
        - name: to_node
          in: query
          description: Filter by destination node number
          required: false
          schema:
            type: integer
        - name: channel
          in: query
          description: Filter by channel number
          required: false
          schema:
            type: integer
        - name: encrypted
          in: query
          description: Filter by encryption status (true for encrypted only, false for decrypted only)
          required: false
          schema:
            type: boolean
        - name: since
          in: query
          description: Unix timestamp to filter packets after this time
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Packet logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PacketListResponse'
        '400':
          description: Bad request (invalid parameters)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /packets/{id}:
    get:
      tags:
        - Packets
      summary: Get a specific packet
      description: Retrieve a specific packet log entry by its ID
      parameters:
        - name: id
          in: path
          description: Packet log ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Packet retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PacketResponse'
        '400':
          description: Bad request (invalid ID)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Packet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...